requirements.txt
PyQt6>=6.9.0
requests>=2.32.3
pygments>=2.19.1
pyperclip>=1.9.0
qt-material>=2.14
pillow>=11.2.1
pyinstaller>=6.13.0 